Estructura completa de componentes y validaciones
Incluye home page, login, registro de usuario y voluntario, validaciones, gráficos funcionales, y frase diaria.
