+++ package.json
@@ -11,8 +11,13 @@
   },
   "dependencies": {
     "axios": "^1.9.0",
+    "chart.js": "^4.5.0",
+    "date-fns": "^4.1.0",
     "react": "^19.1.0",
+    "react-chartjs-2": "^5.3.0",
+    "react-datepicker": "^8.4.0",
     "react-dom": "^19.1.0",
+    "react-icons": "^5.5.0",
     "react-router-dom": "^7.6.2"
   },
   "devDependencies": {
