+++ src/pages/Calendar/Calendar.jsx
@@ -1,0 +1,4 @@
+function Calendar() {
+    return <h1>Calendar</h1>
+}
+export default Calendar
