+++ src/pages/Chat/Emergenci.jsx
@@ -1,0 +1,5 @@
+function Emergenci() {
+    return <h1>Chat de emergencia</h1>
+}
+
+export default Emergenci
