--- src/pages/Home.jsx
@@ -1,13 +1,0 @@
-import { useState } from 'react';
-
-function Home() {
-
-  return (
-    <div>
-      <h1>Bienvenido a EQUILIBRIUM 🧘‍♀️</h1>
-      <p>Estado del backend: Nítido. Soy Joel el Dinosaurio 🦖... </p>
-    </div>
-  );
-}
-
-export default Home;
