--- src/pages/MainUser.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-
-export default function MainUser() {
-  return (
-    <div>
-      <h2>Página Principal Usuario</h2>
-      <p>Bienvenido, usuario.</p>
-    </div>
-  );
-}
