--- src/pages/MainVolunteer.jsx
@@ -1,10 +1,0 @@
-import React from "react";
-
-export default function MainVolunteer() {
-  return (
-    <div>
-      <h2>Página Principal Voluntario</h2>
-      <p>Bienvenido, voluntario.</p>
-    </div>
-  );
-}
