+++ src/pages/Notification/Notification.jsx
@@ -1,0 +1,4 @@
+function Notification() {
+    return <h1>Notificacion</h1>
+}
+export default Notification
