+++ src/pages/Quotes/Quotes.jsx
@@ -1,0 +1,4 @@
+function Quotes() {
+    return <h1>Citas</h1>
+}
+export default Quotes
