--- src/pages/Resources.jsx
@@ -1,5 +1,0 @@
-function Resources() {
-    return <h1>Recursos Educativos 📚</h1>;
-  }
-  export default Resources;
-
